refactor: 맵 데이터 기반 월드 크기·스폰 위치 결정 로직 파일 분리
